# openapi.yaml
openapi: "3.0.0"
info:
  title: LBRY Video Transcoder
  version: "0.0.1"
servers:
  - url: https://api.example.com/api/v1

paths:
  /video/{type}/{url}:
    get:
      summary: Get a video stream
      responses:
        "200":
          description: transcoded stream found and can be delivered
          content:
            application/x-mpegURL: {}
        "202":
          description: transcoding is underway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscodingProgress"
        "403":
          description: transcoded stream was not found but will not be queued for processing
        "404":
          description: stream not found
      parameters:
      - name: url
        in: path
        required: true
        schema:
          type: string
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
           - dash
           - hls
           - range
      - name: touch
        in: query
        required: false
        description: >
          if present in the query, will only return http/transcoding status,
          omitting video data even if present
        schema:
          type: boolean
          default: false

components:
  schemas:
    URL:
      description: LBRY content URL
      type: string
      pattern: ^lbry://.+
    TranscodingProgress:
      type: object
      properties:
        progress:
          type: integer
          minimum: 0
          maximum: 100
        speed:
          type: number
          minimum: 0
        started:
          type: string
          format: date-time
    TranscodingTask:
      type: object
      required:
        - claim_id
        - signed_channel_address
      properties:
        url:
          $ref: "#/components/schemas/URL"
        claim_id:
          type: string
          format: byte
          maxLength: 41
        progress:
         $ref: "#/components/schemas/TranscodingProgress"
        encoding_parameters:
          nullable: true
          type: object
          properties:
            resolution:
              type: array
              enum:
                - 1080p
                - 720p
                - 360p
        created:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - waiting
            - abandoned
            - encoding
            - done
        signed_channel_address:
          type: string
          format: byte
          description: |
            Signed channel address should be a channel address signed
            by channel's key owner.
            Transcoder will validate this signature against channel's public key 
            to prevent unauthorized requests.